tests/wpt/web-platform-tests/eventsource/request-accept.htm

<!DOCTYPE html>
<html>
  <head>
    <title>EventSource: Accept header</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      test.step(function() {
        var source = new EventSource("resources/accept.event_stream?pipe=sub")
        source.onmessage = function(e) {
          test.step(function() {
            assert_equals(e.data, "text/event-stream")
            this.close()
          }, this)
          test.done()
        }
      })
    </script>
  </body>
</html>
